split up table and db funcs
